Swap out __dirname for process.cwd()
Wrote tests for destroy function. Added to prototype

Need to fix destroy, unlink might not be the best method

Added destroy method for the filesystem with tests

Removed package.json from commit

Destroy for filebin is implemented with tests. Fixed catch(done) on test

Removed mocha as a dependency sinces its already in the package.json as dev dep